Add garbage-variant dirty-eligibility tracking; extend indirect_dispatch for value-keyed/flag-changing PHA-RTS tables
Dirty-variant/prune tracking (checkpoint):
- v2_regen.py: a variant the garbage-variant detector (ar_garbage_variant_trap,
  a proven wrong-width decode) flags now counts as prune-eligible for the
  emit-truth pruner when a clean canonical sibling exists, via a separate
  _DIRTY_ONLY_MARKERS list (kept out of _lint_stubs's hard-failure set,
  since an unprunable garbage variant is expected to survive as loud
  runtime-trap residue rather than fail the build).
- codegen.py/emit_function.py/emit_bank.py: set_canonical_variants /
  set_proven_equivalent plumbing for the above, plus related emit-truth
  bookkeeping.
- common_rtl.h/cpu_state.h/cpu_trace.h/cpu_state.c/common_cpu_infra.c:
  supporting runtime hooks and trap infrastructure.

New this pass — extending indirect_dispatch to handle more PHA/RTS idioms:
- indirect_dispatch: add idx:A (switch keyed on the PHA'd table word instead
  of a runtime index register) and sep:<mask> (apply a SEP before dispatch).
  Needed for PHA/RTS tables where the index register gets clobbered/restored
  (e.g. wrapped in PHX/PLX) between the table read and the dispatch, or where
  the dispatch site itself changes the m-flag.
- cfg.py _identify_leaders: promote successors of ANY multi-successor insn to
  block leaders, not just hardcoded block-ender mnemonics. A PHA-stamped
  indirect_dispatch's `ret:` continuation wasn't getting a C label, producing
  "undeclared label" build errors.
- cpu_state.c: generalize the object-loop BRA/BRL-follow into a full RTS/RTL
  chain-follow on dispatch miss (pop, retry lookup, hop-capped at 8). Some
  dispatch idioms pass their continuation through bare RTS/RTL stubs that are
  never real function entries; registering each one individually is unbounded.
- cpu_trace.h/common_cpu_infra.c: cpu_trace_dispatch_oob was a silent no-op in
  non-trace builds, so an indirect_dispatch site misindexing produced zero
  diagnostic output. Wired to an always-on, deduped, capped tripwire
  ([dispatch-oob]) so a bad cfg count or wrong idx model surfaces immediately
  instead of silently degrading.
- snes_other.c: clamp ROM-header-candidate shift counts (UBSan left-shift UB
  on garbage probe bytes; harmless in practice, just noisy under sanitizers).

+++ runner/src/snes/dsp.c
@@ -412,10 +412,15 @@ static void dsp_decodeBrr(Dsp* dsp, int ch) {
       s = curByte >> 4;
     }
     if(s > 7) s -= 16;
+    /* s can be negative here (BRR nibble sign-extended to [-8,7]); left-
+     * shifting a negative signed int is UB in C (UBSan: "left shift of
+     * negative value", 2026-06-30). Hardware just moves bits regardless of
+     * sign, so shift the unsigned reinterpretation and cast back -- same
+     * bit pattern, well-defined. */
     if(shift <= 0xc) {
-      s = (s << shift) >> 1;
+      s = (int)((unsigned)s << shift) >> 1;
     } else {
-      s = (s >> 3) << 12;
+      s = (int)((unsigned)(s >> 3) << 12);
     }
     switch(filter) {
       case 1: s += old + (-old >> 4); break;
